+++ packages/erb_infinite_scroll/lib/src/const/erb_paged_setting.dart
@@ -1,0 +1,10 @@
+class ERbPagedSetting {
+  static const int limit = 20;
+  static const bool pullToRefresh = false;
+  static const bool useSliver = false;
+  static const bool enableInfiniteScroll = true;
+  static const bool shrinkWrap = false;
+  static const bool reverse = false;
+  static const int numSkeletons = 10;
+  static const int firstPageKey = 0;
+}